+++ packages/editor/email-renderer/src/static/__tests__/index.ts
@@ -1,12 +1,11 @@
-import { getImageProcessor, createContentId } from '..';
+import { processImages, createContentId } from '..';
 import * as icons from '../icons';
 
 describe('static asset rendering tests', () => {
-  it('getImageProcessor: should have expected embeddedImages array', () => {
+  it('processImages: should have expected embeddedImages array', () => {
     const contentId = createContentId('info');
     const htmlTestString = `<html><img src="${contentId}" /></html>`;
-    const processor = getImageProcessor(false);
-    const output = processor(htmlTestString);
+    const output = processImages(htmlTestString, false);
 
     // htmlTestString should remain unchanged!
     expect(output.result).toMatch(htmlTestString);
@@ -21,11 +20,10 @@ describe('static asset rendering tests', () => {
     ]);
   });
 
-  it('getImageProcessor: should replace image source with inline source when mock enabled', () => {
+  it('processImages: should replace image source with inline source when mock enabled', () => {
     const contentId = createContentId('info');
     const htmlTestString = `<html><img src="${contentId}" /></html>`;
-    const processor = getImageProcessor(true);
-    const output = processor(htmlTestString);
+    const output = processImages(htmlTestString, true);
 
     // image src should be base64 inline data uri
     expect(output.result).toMatch(`<html><img src="${icons.info}" /></html>`);