+++ packages/search/global-search/src/__tests__/unit/GlobalQuickSearchWrapper.test.tsx
@@ -7,7 +7,6 @@ import {
 
 import { ConfluenceQuickSearchContainer } from '../../components/confluence/ConfluenceQuickSearchContainer';
 import GlobalQuickSearch from '../../components/GlobalQuickSearchWrapper';
-import { HomeQuickSearchContainer } from '../../components/home/HomeQuickSearchContainer';
 import { JiraQuickSearchContainer } from '../../components/jira/JiraQuickSearchContainer';
 import PrefetchedResultsProvider from '../../components/PrefetchedResultsProvider';
 import { mountWithIntl } from './helpers/_intl-enzyme-test-helper';
@@ -18,14 +17,6 @@ jest.mock('../../api/CachingConfluenceClient');
 jest.mock('../../components/AbTestProvider', () => ({
   ABTestProvider: ({ children }: any) => children({}),
 }));
-
-it('should render the home container with context home', () => {
-  const wrapper = mountWithIntl(
-    <GlobalQuickSearch cloudId="123" context="home" />,
-  );
-
-  expect(wrapper.find(HomeQuickSearchContainer).exists()).toBe(true);
-});
 
 it('should render the confluence container with context confluence', () => {
   const wrapper = mountWithIntl(