+++ packages/search/global-search/src/__tests__/unit/util/features.test.tsx
@@ -14,6 +14,7 @@ describe('isInFasterSearchExperiment', () => {
     useUrsForBootstrapping: false,
     disableJiraPreQueryPeopleSearch: false,
     enablePreQueryFromAggregator: false,
+    isAutocompleteEnabled: false,
   };
 
   it('returns false if part of an experiment that is not faster-search', () => {